feat: Add Host header override support for manual DNS resolution
