+++ client/transport/streamable_http_test.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"net/http"
 	"net/http/httptest"
+	"net/url"
 	"strings"
 	"sync"
 	"testing"
@@ -985,4 +986,146 @@ func TestStreamableHTTP_SendNotification_Unauthorized_StaticToken(t *testing.T) 
 	if !errors.Is(err, ErrUnauthorized) {
 		t.Fatalf("Expected ErrUnauthorized, got %T: %v", err, err)
 	}
+}
+
+// TestStreamableHTTPHostOverride tests Host header override for StreamableHTTP transport
+func TestStreamableHTTPHostOverride(t *testing.T) {
+	// Create a test server that captures the Host header
+	var capturedHost string
+	var mu sync.Mutex
+
+	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		mu.Lock()
+		capturedHost = r.Host
+		mu.Unlock()
+
+		// Handle initialize request for StreamableHTTP
+		if r.Method == http.MethodPost {
+			var request JSONRPCRequest
+			decoder := json.NewDecoder(r.Body)
+			if err := decoder.Decode(&request); err != nil {
+				http.Error(w, fmt.Sprintf("Invalid JSON: %v", err), http.StatusBadRequest)
+				return
+			}
+
+			switch request.Method {
+			case "initialize":
+				w.Header().Set("Content-Type", "application/json")
+				w.WriteHeader(http.StatusAccepted)
+				if err := json.NewEncoder(w).Encode(JSONRPCResponse{
+					JSONRPC: request.JSONRPC,
+					ID:      request.ID,
+					Result:  json.RawMessage(`"initialized"`),
+				}); err != nil {
+					http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+					return
+				}
+				return
+			case "test":
+				w.Header().Set("Content-Type", "application/json")
+				w.WriteHeader(http.StatusOK)
+				if err := json.NewEncoder(w).Encode(JSONRPCResponse{
+					JSONRPC: request.JSONRPC,
+					ID:      request.ID,
+					Result:  json.RawMessage(`"test"`),
+				}); err != nil {
+					http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+					return
+				}
+				return
+			}
+		}
+
+		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+	})
+
+	testServer := httptest.NewServer(handler)
+	defer testServer.Close()
+
+	// Parse test server URL to get the actual host
+	serverURL, _ := url.Parse(testServer.URL)
+	actualHost := serverURL.Host
+
+	t.Run("Default Host (no override)", func(t *testing.T) {
+		capturedHost = ""
+		trans, err := NewStreamableHTTP(testServer.URL)
+		require.NoError(t, err)
+		defer trans.Close()
+
+		ctx := context.Background()
+		err = trans.Start(ctx)
+		require.NoError(t, err)
+
+		// Send a request to trigger Host header capture
+		request := JSONRPCRequest{
+			JSONRPC: "2.0",
+			ID:      mcp.NewRequestId(int64(1)),
+			Method:  "test",
+		}
+
+		_, err = trans.SendRequest(ctx, request)
+		require.NoError(t, err)
+
+		// Host should match the actual server host
+		mu.Lock()
+		require.Equal(t, actualHost, capturedHost)
+		mu.Unlock()
+	})
+
+	t.Run("Custom Host override", func(t *testing.T) {
+		capturedHost = ""
+		customHost := "api.example.com"
+
+		trans, err := NewStreamableHTTP(testServer.URL, WithStreamableHTTPHost(customHost))
+		require.NoError(t, err)
+		defer trans.Close()
+
+		ctx := context.Background()
+		err = trans.Start(ctx)
+		require.NoError(t, err)
+
+		// Send a request to trigger Host header capture
+		request := JSONRPCRequest{
+			JSONRPC: "2.0",
+			ID:      mcp.NewRequestId(int64(2)),
+			Method:  "test",
+		}
+
+		_, err = trans.SendRequest(ctx, request)
+		require.NoError(t, err)
+
+		// Host should be the custom host, not the actual server host
+		mu.Lock()
+		require.Equal(t, customHost, capturedHost)
+		require.NotEqual(t, actualHost, capturedHost)
+		mu.Unlock()
+	})
+
+	t.Run("Custom Host with port", func(t *testing.T) {
+		capturedHost = ""
+		customHost := "backend.internal.com:8443"
+
+		trans, err := NewStreamableHTTP(testServer.URL, WithStreamableHTTPHost(customHost))
+		require.NoError(t, err)
+		defer trans.Close()
+
+		ctx := context.Background()
+		err = trans.Start(ctx)
+		require.NoError(t, err)
+
+		// Send a request to trigger Host header capture
+		request := JSONRPCRequest{
+			JSONRPC: "2.0",
+			ID:      mcp.NewRequestId(int64(3)),
+			Method:  "test",
+		}
+
+		_, err = trans.SendRequest(ctx, request)
+		require.NoError(t, err)
+
+		// Host should be the custom host with port
+		mu.Lock()
+		require.Equal(t, customHost, capturedHost)
+		mu.Unlock()
+	})
 }
